refactor: 'utils' -> 'util'
