+++ src/lib/util/index.ts
@@ -1,3 +1,4 @@
+export { default as createStyleContext } from "./createStyleContext";
 export { default as emToPx } from "./emToPx";
 export { default as mapArrayToObject } from "./mapArrayToObject";
 export { default as sleep } from "./sleep";
