+++ src/lib/util/mapArrayToObject.ts
